Fixed q,k,v bug in self attention block. Refactoring.
