+++ src/anyGPT/__init__.py
@@ -1,3 +1,5 @@
 import os.path
 
 DEFAULT_DIR = os.path.join(os.path.expanduser("~"), ".cache", "anygpt")
+DEFAULT_DATADIR = os.path.join(DEFAULT_DIR, "data")
+RAW_DATADIR = os.path.join(DEFAULT_DATADIR, "raw_data")
