+++ src/anyGPT/models/modules.py
@@ -62,6 +62,10 @@ class CausalSelfAttention(nn.Module):
         # compute query, key, values for all heads in batch and move head forward to be the batch dim
         q, k, v = self.c_attn(x).split(self.embedding_size, dim=2)
 
+        k = k.view(B, T, self.num_heads, C // self.num_heads).transpose(1, 2)
+        q = q.view(B, T, self.num_heads, C // self.num_heads).transpose(1, 2)
+        v = v.view(B, T, self.num_heads, C // self.num_heads).transpose(1, 2)
+
         if self.flash:
             y = F.scaled_dot_product_attention(
                 q,
